// server/src/interfaces/vaccination.interfaces.ts
export interface RkiVaccinationReportBundesland {
    Datum: string;
    Bundesland: string;
    BundeslandId_Impfort: string;
    Impfungen_gesamt: string;
    Impfungen_gesamt_min1: string;
    Impfungen_gesamt_voll: string;
    Impfungen_gesamt_boost: string;
    Impfquote_gesamt_min1: string;
    Impfquote_12bis17_min1: string;
    Impfquote_18plus_min1: string;
    Impfquote_18bis59_min1: string;
    Impfquote_60plus_min1: string;
    Impfquote_gesamt_voll: string;
    Impfquote_12bis17_voll: string;
    Impfquote_18plus_voll: string;
    Impfquote_18bis59_voll: string;
    Impfquote_60plus_voll: string;
    Impfquote_gesamt_boost: string;
    Impfquote_12bis17_boost: string;
    Impfquote_18plus_boost: string;
    Impfquote_18bis59_boost: string;
    Impfquote_60plus_boost: string;
}